// connect/src/environments/rabbit.enum.ts
export enum RabbitChannelEnum {
  connect = 'async_events_connect_micro',
  connectFolders = 'async_events_connect_folders_micro',
  connectFoldersExport = 'async_events_connect_folders_export_micro',
}

export enum RabbitExchangesEnum {
  asyncEvents = 'async_events',
  connectFolders = 'connect_folders',
  connectFoldersExport = 'connect_folders_export',
}